refactor: move fake recorder to metrics/test
